+++ frontend/src/App.css
@@ -1,0 +1,10 @@
+.newFeature {
+    border-color: black;
+    border-style: solid;
+    padding: 5px;
+    width: 98.25%;
+}
+
+.contains-counter {
+    padding: 5px;
+}
